api/requirements.txt
flask == 2.3.3
flask-sqlalchemy == 2.5.1
pymysql == 1.1.1
python-dotenv == 1.0.1
sqlalchemy == 1.4.47
marshmallow-sqlalchemy == 0.28.0
flask-jwt-extended == 4.6.0
passlib == 1.7.4
flask-cors == 5.0.0
bcrypt == 4.2.0

# Production WSGI HTTP server 
gunicorn == 20.1.0